Removed Sass
Compiling the Sass assets turned out to be problematic, and since I
planned on getting rid of it anyway, I'm fine with doing it now.

Converting the Sass code to regular CSS3 was _mostly_ easy, since
very few Sass functions where used. The result is not the prettiest
CSS in the world, but it works and can always be refactored later.